get all question api created and user admin api issue resolved

+++ Trello_quora/quora-service/src/main/java/com/upgrad/quora/service/business/AdminBusinessService.java
@@ -11,6 +11,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.ZonedDateTime;
+
 @Service
 public class AdminBusinessService {
     @Autowired
@@ -21,28 +23,31 @@ public class AdminBusinessService {
     @Transactional(propagation = Propagation.REQUIRED)
     public UserEntity getUser(String userUuid, String authorization) throws AuthorizationFailedException, UserNotFoundException {
         UserAuthTokenEntity userAuthTokenEntity = commonDao.getUserAuthToken(authorization);
-        //if (userAuthTokenEntity == null) {
-         //   if (userAuthTokenEntity.getExpiresAt() == null) {
+        if (userAuthTokenEntity == null) {
+            throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
+        } else {
+            ZonedDateTime logOutTime = userAuthTokenEntity.getLogoutAt();
+            ZonedDateTime expiryTime = userAuthTokenEntity.getExpiresAt();
+            ZonedDateTime currentTime = ZonedDateTime.now();
+            if ((logOutTime == null || expiryTime == null) || (!(logOutTime.isBefore(currentTime) || expiryTime.isBefore(currentTime)))) {
                 UserEntity userEntity = commonDao.getUser(userUuid);
                 if (userEntity != null) {
                     String role = userEntity.getRole();
-                    System.out.println("Role is"+role);
-                  //  if (!role.equalsIgnoreCase("nonadmin")) {
-                        System.out.println("inside role"+userUuid);
-                         commonDao.deleteUser(userUuid);
-                        return  userEntity;
-//                    } else {
-//                        throw new AuthorizationFailedException("ATHR-003", "Unauthorized Access, Entered user is not an admin");
-//                    }
+                    if (!role.equalsIgnoreCase("nonadmin")) {
+                        commonDao.deleteUser(userUuid);
+                        return userEntity;
+                    } else {
+                        throw new AuthorizationFailedException("ATHR-003", "Unauthorized Access, Entered user is not an admin");
+                    }
                 } else {
                     throw new UserNotFoundException("USR-001", "User with entered uuid does not exist'.");
                 }
-//            } else {
-//                throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to get user details");
-//            }
-//        } else {
-//            throw new AuthorizationFailedException("ATHR-001", "User is not Signed in, sign in first to upload an image");
-//        }
+            } else {
+                throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to get user details");
+            }
+        }
     }
 }
+
+
